pyproject.toml
[project]
name = "chatbotmedico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flashrank>=0.2.10",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.0",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
]
